examples/cv_satellite_country_prediction/requirements.txt
torch==1.8.1
torchvision==0.9.1
Pillow==8.2.0
pandas==1.2.1
numpy==1.19.5
boto3==1.17.69
scikit-learn==0.24.2
